changed app.js into pages with react router dom. Now we can set multiple pages and will convert server endpoints to routes to clean up and organize.
