+++ client/gui/src/components/homepage/SideImage.jsx
@@ -1,0 +1,10 @@
+//import connect from '../../images/connect.jpg'
+import landing from '../../images/landing.png'
+
+export const SideImage = () =>{
+    return (
+        <div>
+            <img src={landing} alt='landing-pic'/>
+        </div>
+    )
+}
