+++ client/gui/src/components/navigation/navbar.jsx
@@ -1,4 +1,5 @@
 import logo from '../../images/logo.png'
+import {Button} from "@mui/material"
 
 export const Navbar = () =>{
 
@@ -6,6 +7,7 @@ export const Navbar = () =>{
         <nav>
             <img src ={logo}></img>
             <ul className="list-container">
+                <li><Button variant = "contained" color ="secondary">Contact</Button></li>
             </ul>
         </nav>
     )
