--- client/gui/src/components/signup/SideImage.jsx
@@ -1,10 +1,0 @@
-//import connect from '../../images/connect.jpg'
-import smol from '../../images/smol.png'
-
-export const SideImage = () =>{
-    return (
-        <div>
-            <img src={smol} alt='landing-pic'/>
-        </div>
-    )
-}
